refactor: improve rust
